fix bug in receive logic since net.Addr is not comparbale
net.Addr is an interface so a naive == comparison will compare the
interface values and not the actual IP/port values.

Comparing the `.String()` form of the addresses is not the best either
because it assumes that every net.Addr will generate a `.String()` result
that is always IP:port.