enable parallel test execution
Note: The parallel tests must be run inside a subtest so that the test TFTP
servers are not closed before the tests have run. Previously, when running
the tests in parallel inside the outer tests (e.g., TestReadSupport and
TestWriteSupport), the test servers were shutdown prematurely and caused
the tests to hit their timeouts.